+++ app/src/main/java/com/kickstarter/libs/htmlparser/HTMLModels.kt
@@ -1,0 +1,76 @@
+package com.kickstarter.libs.htmlparser
+
+import org.jsoup.nodes.Element
+
+interface ViewElement
+
+data class VideoViewElement(val sourceUrls: List<String>) : ViewElement
+data class TextViewElement(var components: List<TextComponent>) : ViewElement {
+    // TODO: this should probably be a spannableString
+    val attributedText: String
+        get() {
+            var string = ""
+            for (component in components) {
+                string += component.text
+            }
+            return string
+        }
+}
+
+data class TextComponent(val text: String, val link: String?, val styles: List<TextStyleType>) {
+    enum class TextStyleType(val tag: String?) {
+        PARAGRAPH("p"),
+        HEADER1("h1"),
+        BOLD("strong"),
+        LINK("a"),
+        EMPHASIS("em"),
+        CAPTION("figcaption"),
+        LIST("li"),
+        UNKNOWN(null);
+
+        companion object {
+            fun initialize(tag: String): TextStyleType {
+                return (values().find { it.tag == tag }) ?: UNKNOWN
+            }
+        }
+    }
+}
+
+data class EmbeddedLinkViewElement(
+    val href: String,
+    override val src: String,
+    val caption: String?
+) : ImageViewElement(src)
+
+open class ImageViewElement(open val src: String) : ViewElement
+
+enum class ViewElementType(val tag: String?) {
+    IMAGE("img"),
+    TEXT(null),
+    VIDEO("video"),
+    EMBEDDED_LINK(null),
+    OEMBED(null),
+    UNKNOWN(null);
+
+    companion object {
+        fun initialize(element: Element): ViewElementType {
+            val tag = element.tag().name
+            if (tag == "div") {
+                for (attribute in element.attributes()) {
+                    if (attribute.key == "class" && attribute.value == "template oembed") {
+                        return OEMBED
+                    }
+                }
+            } else if (TextComponent.TextStyleType.initialize(tag) == TextComponent.TextStyleType.LINK && !element.getElementsByTag("img").isEmpty()) {
+                return EMBEDDED_LINK
+            } else if (TextComponent.TextStyleType.values().map { it.tag }.contains(tag)) {
+                return TEXT
+            } else if (tag == IMAGE.tag) {
+                return IMAGE
+            } else if (tag == VIDEO.tag) {
+                return VIDEO
+            }
+            return UNKNOWN
+        }
+    }
+}
