+++ app/src/main/java/com/kickstarter/libs/htmlparser/HTMLParser.kt
@@ -1,82 +1,9 @@
-package com.kickstarter.libs
+package com.kickstarter.libs.htmlparser
 
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
 import org.jsoup.nodes.TextNode
 import org.jsoup.select.Elements
-
-interface ViewElement
-
-data class VideoViewElement(val sourceUrls: ArrayList<String>) : ViewElement
-data class TextViewElement(var components: List<TextComponent>) : ViewElement {
-
-    val attributedText: String
-        get() {
-            var string = ""
-            for (component in components) {
-                string += component.text
-            }
-            return string
-        }
-}
-
-data class TextComponent(val text: String, val link: String?, val styles: List<TextStyleType>)
-
-data class EmbeddedLinkViewElement(
-    val href: String,
-    override val src: String,
-    val caption: String?
-) : ImageViewElement(src)
-
-open class ImageViewElement(open val src: String) : ViewElement
-
-enum class TextStyleType(val tag: String?) {
-    PARAGRAPH("p"),
-    HEADER1("h1"),
-    BOLD("strong"),
-    LINK("a"),
-    EMPHASIS("em"),
-    CAPTION("figcaption"),
-    LIST("li"),
-    UNKNOWN(null);
-
-    companion object {
-        fun initialize(tag: String): TextStyleType {
-            return (values().find { it.tag == tag }) ?: UNKNOWN
-        }
-    }
-}
-
-enum class ViewElementType(val tag: String?) {
-    IMAGE("img"),
-    TEXT(null),
-    VIDEO("video"),
-    EMBEDDED_LINK(null),
-    OEMBED(null),
-    UNKNOWN(null);
-
-    companion object {
-        fun initialize(element: Element): ViewElementType {
-            val tag = element.tag().name
-            if (tag == "div") {
-                for (attribute in element.attributes()) {
-                    if (attribute.key == "class" && attribute.value == "template oembed") {
-                        return OEMBED
-                    }
-                }
-            } else if (TextStyleType.initialize(tag) == TextStyleType.LINK && !element.getElementsByTag("img").isEmpty()) {
-                return EMBEDDED_LINK
-            } else if (TextStyleType.values().map { it.tag }.contains(tag)) {
-                return TEXT
-            } else if (tag == IMAGE.tag) {
-                return IMAGE
-            } else if (tag == VIDEO.tag) {
-                return VIDEO
-            }
-            return UNKNOWN
-        }
-    }
-}
 
 class HTMLParser {
 
@@ -144,7 +71,7 @@ class HTMLParser {
         for (node in element.childNodes()) {
             (node as? TextNode)?.let {
                 val href = (element.attributes().firstOrNull { it.key == "href" })?.value
-                val textStyleList = tags.map { tag -> TextStyleType.initialize(tag) }
+                val textStyleList = tags.map { tag -> TextComponent.TextStyleType.initialize(tag) }
                 textComponents.add(TextComponent(it.text(), href, ArrayList(textStyleList)))
             }
             (node as? Element)?.let {
